Update README.md
This project was inspired by my best friend and her wonderful boys who are driven, motivated, and ready to be entrepreneurs. The aim of this project is to empower other kids like them to step into the entrepreneurial world. It provides guidance for them to start their own business, allowing them to sell items they no longer need as well as their original artwork, pottery, photography, and more.